requirements.txt
tensorflow
numpy
matplotlib
pandas
seaborn
PyYAML

## local package installation
-e .  ## <<< to install local packages defined in setup.py